+++ packages/client/src/components/AlgorithmSolver/algorithm.js
