+++ packages/client/src/pages/Algorithm/DijkstraComponent/DijkstraGraph.jsx
@@ -2,6 +2,8 @@ import React, { useState, useEffect } from "react";
 import * as d3 from "d3";
 import { Button, Stack } from "@mui/material";
 import Graph from "../../../components/solver/algorithm.js";
+import Animation from "../../../components/GraphRenderer/animate";
+
 
 export const DijkstraGraph = () => {
   const [nodes, setNodes] = useState([
